+++ szamlaagent/src/SzamlaAgent/Item/ReceiptItem.php
@@ -70,6 +70,10 @@ class ReceiptItem extends Item {
             $data['fokonyv']      = $this->getLedgerData()->buildXmlData();
         }
 
+        if (SzamlaAgentUtil::isNotNull($this->getDataDeletionCode())) {
+            $data['torloKod'] = SzamlaAgentUtil::nonNegativeInteger($this->getDataDeletionCode());
+        }
+
         return $data;
     }