+++ szamlaagent/src/SzamlaAgent/SzamlaAgentRequest.php
@@ -462,6 +462,12 @@ class SzamlaAgentRequest {
 
             $ch = curl_init($agent->getApiUrl());
 
+            if ($agent->hasCertification()) {
+                curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, true);
+                curl_setopt($ch, CURLOPT_SSL_VERIFYHOST, 2);
+                curl_setopt($ch, CURLOPT_CAINFO, $agent->getCertificationFilePath());
+            }
+
             curl_setopt($ch, CURLOPT_POST, true);
             curl_setopt($ch, CURLOPT_HEADER, true);
             curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);